Change from azure functions to dockerfile
