+++ backend_azure_function/GenerateQuiz/requirements.txt
@@ -1,0 +1,7 @@
+# Do not include azure-functions-worker in this file
+# The Python Worker is managed by the Azure Functions platform
+# Manually managing azure-functions-worker may cause unexpected issues
+
+openai
+fastapi
+uvicorn
